refactor: split transform corrections
