+++ scripts/transforms/__init__.py
@@ -1,0 +1,1 @@
+"""Transform implementations grouped by responsibility."""
